♻️ refactor(shared): move the browser-store diagnostics out of data/local/db
